pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uh-grad-planner"
version = "0.2.0"
description = "UH graduation schedule planner - hackathon MVP with Gemini integration"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "UH CodeRED Team", email = "team@example.com" }]
dependencies = [
  "pydantic>=2.8",
  "pdfplumber>=0.11",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
  "requests>=2.32",
  "networkx>=3.3",
  "google-genai>=0.3",          # Gemini API SDK
  "fastapi>=0.115",             # Web API framework
  "uvicorn[standard]>=0.32",    # ASGI server
]

[project.optional-dependencies]
dev = ["pytest>=8.2", "ruff>=0.6"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
